Valider et supprimer service admin

+++ src/web/DeleteServiceController.java
@@ -2,8 +2,10 @@ package web;
 
 import data.Location;
 import data.Service;
+import data.User;
 import ejb.ServiceBean;
 import ejb.TransactionException;
+import ejb.UserBean;
 
 import javax.ejb.EJB;
 import javax.servlet.ServletException;
@@ -15,12 +17,24 @@ import java.io.IOException;
 
 @WebServlet("/delete-service")
 public class DeleteServiceController extends HttpServlet {
+    @EJB
+    UserBean userBean;
 
     @EJB
     ServiceBean serviceBean;
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        serviceBean.deleteById(Long.parseLong(req.getParameter("service_id_delete")));
+
+        User user = Utils.getUser(req,userBean);
+        if (user.isAdmin())
+        {
+            serviceBean.adminDelete(Long.parseLong(req.getParameter("service_id_delete")));
+        }else{
+
+            serviceBean.deleteById(Long.parseLong(req.getParameter("service_id_delete")));
+
+        }
+
         resp.sendRedirect("Servlet");
     }
 }
